split change module
