+++ .gitignore
@@ -6,3 +6,4 @@ deploy
 docker/backend_db/data
 /benchmarks/target
 .idea
+.env
