+++ src/app/endpoint/change/tests/mod.rs
@@ -1,0 +1,6 @@
+mod create;
+pub use self::create::*;
+mod list;
+pub use self::list::*;
+mod delete;
+pub use self::delete::*;
